Revert "Modify announcement topic to handle when changing sizes or finishing"

+++ srcCpp03/perftest_cpp.h
@@ -86,7 +86,7 @@ class perftest_cpp
     int  _LatencyCount;
     int  _NumSubscribers;
     int  _NumPublishers;
-    unsigned long _InstanceCount;
+    int _InstanceCount;
     IMessaging *_MessagingImpl;
     char **_MessagingArgv;
     int _MessagingArgc;
